Move hgrid data to apps
