+++ addons/github/apps.py
@@ -1,7 +1,94 @@
+import logging
 import os
-from addons.base.apps import BaseAddonAppConfig, generic_root_folder
 
-github_hgrid_data = generic_root_folder('github')
+from addons.base.apps import BaseAddonAppConfig, generic_root_folder
+from addons.github.api import GitHubClient, ref_to_params
+from addons.github.exceptions import NotFoundError, GitHubError
+from addons.github.serializer import GitHubSerializer
+from addons.github.utils import get_refs, check_permissions
+from website.util import rubeus
+
+logger = logging.getLogger(__name__)
+
+logging.getLogger('github3').setLevel(logging.WARNING)
+logging.getLogger('requests.packages.urllib3.connectionpool').setLevel(logging.WARNING)
+
+def github_hgrid_data(node_settings, auth, **kwargs):
+
+    # Quit if no repo linked
+    if not node_settings.complete:
+        return
+
+    connection = GitHubClient(external_account=node_settings.external_account)
+
+    # Initialize repo here in the event that it is set in the privacy check
+    # below. This potentially saves an API call in _check_permissions, below.
+    repo = None
+
+    # Quit if privacy mismatch and not contributor
+    node = node_settings.owner
+    if node.is_public and not node.is_contributor(auth.user):
+        try:
+            repo = connection.repo(node_settings.user, node_settings.repo)
+        except NotFoundError:
+            # TODO: Test me @jmcarp
+            # TODO: Add warning message
+            logger.error('Could not access GitHub repo')
+            return None
+        if repo.private:
+            return None
+
+    try:
+        branch, sha, branches = get_refs(
+            node_settings,
+            branch=kwargs.get('branch'),
+            sha=kwargs.get('sha'),
+            connection=connection,
+        )
+    except (NotFoundError, GitHubError):
+        # TODO: Show an alert or change GitHub configuration?
+        logger.error('GitHub repo not found')
+        return
+
+    if branch is not None:
+        ref = ref_to_params(branch, sha)
+        can_edit = check_permissions(
+            node_settings, auth, connection, branch, sha, repo=repo,
+        )
+    else:
+        ref = None
+        can_edit = False
+
+    name_tpl = '{user}/{repo}'.format(
+        user=node_settings.user, repo=node_settings.repo
+    )
+
+    permissions = {
+        'edit': can_edit,
+        'view': True,
+        'private': node_settings.is_private
+    }
+    urls = {
+        'upload': node_settings.owner.api_url + 'github/file/' + (ref or ''),
+        'fetch': node_settings.owner.api_url + 'github/hgrid/' + (ref or ''),
+        'branch': node_settings.owner.api_url + 'github/hgrid/root/',
+        'zip': node_settings.owner.api_url + 'github/zipball/' + (ref or ''),
+        'repo': 'https://github.com/{0}/{1}/tree/{2}'.format(node_settings.user, node_settings.repo, branch)
+    }
+
+    branch_names = [each.name for each in branches]
+    if not branch_names:
+        branch_names = [branch]  # if repo un-init-ed then still add default branch to list of branches
+
+    return [rubeus.build_addon_root(
+        node_settings,
+        name_tpl,
+        urls=urls,
+        permissions=permissions,
+        branches=branch_names,
+        defaultBranch=branch,
+        private_key=kwargs.get('view_only', None),
+    )]
 
 HERE = os.path.dirname(os.path.abspath(__file__))
 NODE_SETTINGS_TEMPLATE = os.path.join(
